feat(gcb): Add endpoint to fetch build artifacts (#433)
* feat(gcb): Add support for parsing GCS paths

In order to fetch the list of artifacts produced by a build, we'll
need to fetch the manifest from GCS. This will involve translating
a GCS path into a bucket/object/version, which there is no helper
function in the GCS library for; write a small class to do this and
test it.

* feat(gcb): Add GCS object to GCB account

As Google Cloud Build stores a manifest of produced artifacts in
Google Cloud Storage, a GCB account will need to be able to talk
to GCS in order to return information about produced artifacts.

This allows us to better encapsulate information about produced artifacts
rather than having the GCB account simply return a path to the manfest
and require the caller to parse any artifacts (which are in a GCB-specific
format).

* feat(gcb): Add endpoint to fetch build artifacts

The endpoint should return a list of Spinnaker artifacts representing
artifacts produced by the build.  A build can produce either docker
containers or arbitrary files in Google Cloud Storage.

Docker containers are directly included in the build output, so we
can read the relevant fields and construct an artifact. GCS files
are not included directly in the output; the output contains a
GCS path to a manifest of produced files. We need to fetch and
parse this manifest in order to generate the produced artifacts.

* feat(gcb): Add endpoint for extracting artifacts from build

In some cases the consumer already has a build from which they'd like
to extract artifacts; allow the consumer to send this build to igor
rather than just the id (which could have igor fetch stale data).