fix(travis): Correctly parse global_env from Travis v3 API (#504)
When the `config` section is returned as part of a *build* (which is unsupported in the Travis v3 API), `global_env` is a list of Strings. When it is returned as part of a *job*, `global_env` is instead a space separated String. And since env variable values can also contain spaces and equal signs, I had to create a little regex parser thingy.
This also fixes an issue in Deck where you can't add parameters to a Travis stage if `.travis.yml` contains anything in the `env.global` section.