refactor(gcb): Reduce visibility of GCB packages (#593)
* refactor(gcb): Move all GCB files into the same package

The GoolgeCloudBuildArtifact and GoolgeCloudBuildConfig classes
were in different packages; by moving these to the same package
as the other GCB classes we'll be able to reduce the visibility
of many GCB classes to package-private.

* refactor(gcb): Make packages package-private and final where possible

The only external contract the com.netflix.spinnaker.igor.gcb
package intends to provide is the GoogleCloudBuildController
component; everything else is an implementation detail.

To better communicate and enforce this:
* Make the controller (and all mapped methods) public; some of
these were package-private before which works because Spring uses
reflection to call controllers, but we should document them
accurately as being called outside of the package.
* Make all other classes (and methods) that were public package-private

Finally (no pun intended), none of the classes in this package
were designed for inheritance so mark them final wherever possible.
The cases where it is not possible are:
* Classes that are mocked in tests as a final class cannot be
mocked; in this case, we'll rely on the fact that the class is
package-private so at least can't be subclassed outside the package.
* Spring controller and configuration classes aren't allowed to be
final

* refactor(gcb): Add nullity annotations

Add default Nonnull to all methods, parameters and fields across the
package, and annotate the few exceptions with Nullable.

This does surface a few places where we could consider replacing a
null with either an Optional or an appropriate null object, but
this PR does not tackle any of those refactors.

* refactor(gcb): A few minor refactors

* Replace statement lambda with expression lambda
* Make GoogleCloudBuildExecutor nicer to clients by adding type
bounds.
* Clean up the null checks of build status by using a null-safe
valueOf that returns UNKNOWN. (This keeps the existing behavior
as UNKNOWN status has the lowest priority, so null status continues
to have the lowest priority).
* Use try-with-resources on JsonGenerator so that it always gets
closed before we use the underlying writer, avoiding the need to
explicitly flush.

* refactor(gcb): Make GCB account repository immutable

As we never add/remove accounts after startup, the repository
should be immutable. Maybe at some point if we support dynamic
accounts we should change this (but at that point we'd need to
think about thread safety anyway).

This commit also makes a few changes to the tests on that class;
they had to be updated anyway to use the new Builder class, and
while I was there I improved the assertions to make them more clear.

* refactor(gcb): Update GCB controller to return immutable lists

Return immutable lists from the controller, and convert dependent
code to also use immutable lists.

* refactor(gcb): Replace RuntimeException with UncheckedIOException

In a number of places in the GCB package, we wrap an IOException in
a RuntimeException; replace this with an UncheckedIOException.