refactor(gcb): Further clean-up of GCB config (#742)
Given that we are enforcing a non-null account name and project
downstream, let's enforce them at object creation to make things
cleaner.

This commit also removes the subscriptionName field; it is
part of the config but not needed by igor, so there's no sense
mapping it.

Finally, make jsonKey an optional to better reflect that it is
optional (instead of treating an empty string as a sentinel
value).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>