chore(travis): Move Travis integration to igor-monitor-travis (#770)
* chore(travis): Move Travis integration to igor-monitor-travis

Had to move some classes to igor-core to be able to do this, as both Jenkins and Travis (maybe other ci's as well?) depends on them. We could probably consider moving some of this to kork instead.
Also converted a bunch of classes to Java. No Groovy left in igor-core (except tests) \o/

* Revert copyright changes

* Remove hystrix